Non-Iterable but traversable Graph
This adds finding an item in the linked list portion of the graph, makes
bidirectional connections, makes only one addition per unique item in
each input line. The Roads class was the tester class used to verify the
functionality. tempInput was the input file made from the handout in
class.