RoadsUI (and some style changes)
I've kept the originals in a "Copy of Originals" folder to avoid causing problems to the main program.

I added findRouteScreen() but Java keeps crashing when I try to use the ComboBoxes--I'm unsure if this is just my computer or the code itself. If you just want to look at it, go through the program to that screen and just go back, or close the program, to avoid it crashing.