Jonathan
I created the minimum spanning tree input screen and a flashing text animation to be used for when the user does not enter any information. I also updated the node and graph classes to allow for Doubles and the alert and confirm classes to have their windows always appear in the center of the main window.