Jonathan (#23)
* Updates

I created the minimum spanning tree input screen and a flashing text animation to be used for when the user does not enter any information. I also updated the node and graph classes to allow for Doubles and the alert and confirm classes to have their windows always appear in the center of the main window.

* GUI Update

I implemented Candace's shortest path input screen and started on the screen to display the results of the minimum spanning tree.